requirements.txt
Pillow 
numpy